oracle-netsuite-faq-more-custom-form-detail
In the custom form section, added detail relating to the fact that the standard forms themselves can be modified. This is probably our most common cause of push failure and we need to make it clear that customised standard forms are incompatible with our system. 

Also removed the line about updating the bundle. Having the latest bundle is relevant for all troubleshooting of the NetSuite integration. Perhaps this could be called out elsewhere in the docs?